+++ packages/components/lib/components/primitives/form/TextField/TextField.d.ts
@@ -4,6 +4,15 @@ export declare enum InputShapes {
     SemiRounded = "rounded-input-semi-rounded",
     Rectangle = "rounded-input-rectangle"
 }
+export declare enum InputMode {
+    Text = "text",
+    Decimal = "decimal",
+    Numeric = "numeric",
+    Tel = "tel",
+    Search = "search",
+    Email = "email",
+    Url = "url"
+}
 declare type Props = {
     className?: string;
     shape?: InputShapes;
@@ -16,6 +25,7 @@ declare type Props = {
     type?: string;
     maxLength?: number;
     onChange?: (event: ChangeEvent<HTMLInputElement>) => void;
+    inputMode?: InputMode;
 };
 export declare const TextField: React.ForwardRefExoticComponent<Props & React.RefAttributes<HTMLInputElement>>;
 export {};