+++ packages/components/lib/components/primitives/form/TextFieldWithLabel/TextFieldWithLabel.stories.js
@@ -32,7 +32,8 @@ const base = () => /*#__PURE__*/_react.default.createElement(_.TextFieldWithLabe
   required: (0, _addonKnobs.boolean)('Required', false),
   shape: (0, _addonKnobs.select)('Shape', _TextField.InputShapes, _TextField.InputShapes.SemiRounded),
   onChange: (0, _addonActions.action)('on change'),
-  maxLength: (0, _addonKnobs.number)('Max length', 33)
+  maxLength: (0, _addonKnobs.number)('Max length', 33),
+  inputMode: (0, _addonKnobs.select)('Input mode', _TextField.InputMode, _TextField.InputMode.Numeric)
 });
 
 exports.base = base;