+++ packages/contentful/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Change Log
 
+## 0.0.26
+
+### Patch Changes
+
+- a96b018: Removed copy from Contentful README
+
+## 0.0.25
+
+### Patch Changes
+
+- Updated dependencies [d28706e]
+  - @cmpsr/components@0.0.20
+
+## 0.0.24
+
+### Patch Changes
+
+- Updated dependencies [8af560d]
+  - @cmpsr/components@0.0.19
+
 ## 0.0.23
 
 ### Patch Changes